+++ src/WebJobs.Extensions.DurableTask/HttpRetryOptions.cs
@@ -2,6 +2,8 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
+using System.Collections.Generic;
+using System.Net;
 using DurableTaskCore = DurableTask.Core;
 
 namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
@@ -9,25 +11,25 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
     /// <summary>
     /// Defines retry policies that can be passed as parameters to various operations.
     /// </summary>
-    public class RetryOptions
+    public class HttpRetryOptions
     {
-        private readonly DurableTaskCore.RetryOptions retryOptions;
+        private readonly DurableTaskCore.RetryOptions coreRetryOptions;
 
-        // Would like to make this durability provider specific, but since this is a customer
+        // Would like to make this durability provider specific, but since this is a developer
         // facing type, that is difficult.
         private static readonly TimeSpan DefaultMaxRetryinterval = TimeSpan.FromDays(6);
 
         /// <summary>
-        /// Creates a new instance RetryOptions with the supplied first retry and max attempts.
+        /// Creates a new instance SerializableRetryOptions with the supplied first retry and max attempts.
         /// </summary>
         /// <param name="firstRetryInterval">Timespan to wait for the first retry.</param>
         /// <param name="maxNumberOfAttempts">Max number of attempts to retry.</param>
         /// <exception cref="ArgumentException">
         /// The <paramref name="firstRetryInterval"/> value must be greater than <see cref="TimeSpan.Zero"/>.
         /// </exception>
-        public RetryOptions(TimeSpan firstRetryInterval, int maxNumberOfAttempts)
+        public HttpRetryOptions(TimeSpan firstRetryInterval, int maxNumberOfAttempts)
         {
-            this.retryOptions = new DurableTaskCore.RetryOptions(firstRetryInterval, maxNumberOfAttempts);
+            this.coreRetryOptions = new DurableTaskCore.RetryOptions(firstRetryInterval, maxNumberOfAttempts);
             this.MaxRetryInterval = DefaultMaxRetryinterval;
         }
 
@@ -39,20 +41,20 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         /// </value>
         public TimeSpan FirstRetryInterval
         {
-            get { return this.retryOptions.FirstRetryInterval; }
-            set { this.retryOptions.FirstRetryInterval = value; }
+            get { return this.coreRetryOptions.FirstRetryInterval; }
+            set { this.coreRetryOptions.FirstRetryInterval = value; }
         }
 
         /// <summary>
         /// Gets or sets the max retry interval.
         /// </summary>
         /// <value>
-        /// The TimeSpan of the max retry interval, defaults to <see cref="TimeSpan.MaxValue"/>.
+        /// The TimeSpan of the max retry interval, defaults to 6 days.
         /// </value>
         public TimeSpan MaxRetryInterval
         {
-            get { return this.retryOptions.MaxRetryInterval; }
-            set { this.retryOptions.MaxRetryInterval = value; }
+            get { return this.coreRetryOptions.MaxRetryInterval; }
+            set { this.coreRetryOptions.MaxRetryInterval = value; }
         }
 
         /// <summary>
@@ -63,8 +65,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         /// </value>
         public double BackoffCoefficient
         {
-            get { return this.retryOptions.BackoffCoefficient; }
-            set { this.retryOptions.BackoffCoefficient = value; }
+            get { return this.coreRetryOptions.BackoffCoefficient; }
+            set { this.coreRetryOptions.BackoffCoefficient = value; }
         }
 
         /// <summary>
@@ -75,8 +77,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         /// </value>
         public TimeSpan RetryTimeout
         {
-            get { return this.retryOptions.RetryTimeout; }
-            set { this.retryOptions.RetryTimeout = value; }
+            get { return this.coreRetryOptions.RetryTimeout; }
+            set { this.coreRetryOptions.RetryTimeout = value; }
         }
 
         /// <summary>
@@ -87,25 +89,26 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         /// </value>
         public int MaxNumberOfAttempts
         {
-            get { return this.retryOptions.MaxNumberOfAttempts; }
-            set { this.retryOptions.MaxNumberOfAttempts = value; }
+            get { return this.coreRetryOptions.MaxNumberOfAttempts; }
+            set { this.coreRetryOptions.MaxNumberOfAttempts = value; }
         }
 
         /// <summary>
-        /// Gets or sets a delegate to call on exception to determine if retries should proceed.
+        /// Gets or sets the list of status codes upon which the
+        /// retry logic specified by this object shall be triggered.
+        /// If none are provided, all 4xx and 5xx status codes
+        /// will be retried.
         /// </summary>
-        /// <value>
-        /// The delegate to handle exception to determine if retries should proceed.
-        /// </value>
-        public Func<Exception, bool> Handle
-        {
-            get { return this.retryOptions.Handle; }
-            set { this.retryOptions.Handle = value; }
-        }
+        public IList<HttpStatusCode> StatusCodesToRetry { get; set; } = new List<HttpStatusCode>();
 
-        internal DurableTaskCore.RetryOptions GetRetryOptions()
+        internal RetryOptions GetRetryOptions()
         {
-            return this.retryOptions;
+            return new RetryOptions(this.FirstRetryInterval, this.MaxNumberOfAttempts)
+                {
+                    BackoffCoefficient = this.BackoffCoefficient,
+                    MaxRetryInterval = this.MaxRetryInterval,
+                    RetryTimeout = this.RetryTimeout,
+                };
         }
     }
 }