+++ test/Common/DurableHttpTests.cs
@@ -8,14 +8,11 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
-using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Identity;
 using Microsoft.Azure.WebJobs.Host.TestCommon;
-using Microsoft.Diagnostics.Tracing;
 using Microsoft.Extensions.Primitives;
 using Moq;
 using Moq.Protected;
@@ -31,6 +28,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
         private readonly ITestOutputHelper output;
 
         private readonly TestLoggerProvider loggerProvider;
+
+        private static int mockSynchronousHttpMessageHandlerCount;
 
         public DurableHttpTests(ITestOutputHelper output)
         {
@@ -113,8 +112,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
       ""tenantid"": ""tenant_id""
     }
   },
-  ""AsynchronousPatternEnabled"": true,
-  ""Timeout"": null
+  ""AsynchronousPatternEnabled"": true
 }";
 
             Dictionary<string, string> headers = new Dictionary<string, string>();
@@ -153,6 +151,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
     }
    },
   ""asynchronousPatternEnabled"": true,
+  ""retryOptions"": null,
   ""timeout"": null
 }";
             ManagedIdentityTokenSource managedIdentityTokenSource = new ManagedIdentityTokenSource("dummy url", options);
@@ -184,6 +183,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
     ""resource"": ""dummy url""
   },
   ""asynchronousPatternEnabled"": true,
+  ""retryOptions"": null,
   ""timeout"": null
 }";
 
@@ -363,6 +363,189 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
         }
 
         /// <summary>
+        /// End-to-end test which checks if the CallHttpAsync Orchestrator retries when the
+        /// HTTP request times out and the CallHttpAsync API throws a TimeoutException.
+        /// </summary>
+        [Theory]
+        [Trait("Category", PlatformSpecificHelpers.TestCategory)]
+        [MemberData(nameof(TestDataGenerator.GetFullFeaturedStorageProviderOptions), MemberType = typeof(TestDataGenerator))]
+        public async Task DurableHttpAsync_Synchronous_RetryExceededWithHttpException(string storageProvider)
+        {
+            mockSynchronousHttpMessageHandlerCount = 0;
+
+            HttpResponseMessage testHttpResponseMessage = CreateTestHttpResponseMessage(HttpStatusCode.OK);
+            HttpMessageHandler httpMessageHandler = MockSynchronousHttpMessageHandlerWithHttpRequestException();
+
+            using (ITestHost host = TestHelpers.GetJobHost(
+                this.loggerProvider,
+                nameof(this.DurableHttpAsync_Synchronous_RetryExceededWithHttpException),
+                enableExtendedSessions: false,
+                storageProviderType: storageProvider,
+                durableHttpMessageHandler: new DurableHttpMessageHandlerFactory(httpMessageHandler)))
+            {
+                await host.StartAsync();
+
+                Dictionary<string, string> headers = new Dictionary<string, string>
+                {
+                    { "Accept", "application/json" },
+                };
+                TestDurableHttpRequest testRequest = new TestDurableHttpRequest(
+                    httpMethod: HttpMethod.Get,
+                    headers: headers,
+                    firstRetryInterval: TimeSpan.FromSeconds(1),
+                    maxNumberOfAttempts: 3);
+
+                string functionName = nameof(TestOrchestrations.CallHttpAsyncOrchestrator);
+                var client = await host.StartOrchestratorAsync(functionName, testRequest, this.output);
+                var status = await client.WaitForCompletionAsync(this.output, timeout: TimeSpan.FromSeconds(400));
+
+                var output = status?.Output;
+                Assert.Equal(3, mockSynchronousHttpMessageHandlerCount);
+                Assert.Contains("No such host is known.", output.ToString());
+                Assert.Equal(OrchestrationRuntimeStatus.Failed, status?.RuntimeStatus);
+
+                await host.StopAsync();
+            }
+        }
+
+        /// <summary>
+        /// End-to-end test which checks if the CallHttpAsync Orchestrator retries when the
+        /// HTTP request times out and the CallHttpAsync API throws a TimeoutException.
+        /// </summary>
+        [Theory]
+        [Trait("Category", PlatformSpecificHelpers.TestCategory)]
+        [MemberData(nameof(TestDataGenerator.GetFullFeaturedStorageProviderOptions), MemberType = typeof(TestDataGenerator))]
+        public async Task DurableHttpAsync_Synchronous_RetryExceededWithHttpNotFoundStatusCode(string storageProvider)
+        {
+            HttpMessageHandler httpMessageHandler = MockSynchronousHttpMessageHandlerWithHttp404();
+
+            using (ITestHost host = TestHelpers.GetJobHost(
+                this.loggerProvider,
+                nameof(this.DurableHttpAsync_Synchronous_RetryExceededWithHttpNotFoundStatusCode),
+                enableExtendedSessions: false,
+                storageProviderType: storageProvider,
+                durableHttpMessageHandler: new DurableHttpMessageHandlerFactory(httpMessageHandler)))
+            {
+                await host.StartAsync();
+
+                Dictionary<string, string> headers = new Dictionary<string, string>
+                {
+                    { "Accept", "application/json" },
+                };
+                TestDurableHttpRequest testRequest = new TestDurableHttpRequest(
+                    httpMethod: HttpMethod.Get,
+                    headers: headers,
+                    firstRetryInterval: TimeSpan.FromSeconds(1),
+                    maxNumberOfAttempts: 3)
+                {
+                    StatusCodesToRetry = new List<HttpStatusCode> { HttpStatusCode.NotFound },
+                };
+
+                string functionName = nameof(TestOrchestrations.CallHttpAsyncOrchestrator);
+                var client = await host.StartOrchestratorAsync(functionName, testRequest, this.output);
+                var status = await client.WaitForCompletionAsync(this.output, timeout: TimeSpan.FromSeconds(400));
+
+                var output = status?.Output;
+                Assert.Equal(3, mockSynchronousHttpMessageHandlerCount);
+                Assert.Equal(OrchestrationRuntimeStatus.Failed, status?.RuntimeStatus);
+
+                await host.StopAsync();
+            }
+        }
+
+        /// <summary>
+        /// End-to-end test which checks if the CallHttpAsync Orchestrator retries when the
+        /// HTTP request times out and the CallHttpAsync API throws a TimeoutException.
+        /// </summary>
+        [Theory]
+        [Trait("Category", PlatformSpecificHelpers.TestCategory)]
+        [MemberData(nameof(TestDataGenerator.GetFullFeaturedStorageProviderOptions), MemberType = typeof(TestDataGenerator))]
+        public async Task DurableHttpAsync_Synchronous_RetryExceededWithHttpNotFoundViaEnsureSuccessException(string storageProvider)
+        {
+            HttpMessageHandler httpMessageHandler = MockSynchronousHttpMessageHandlerWithHttp404();
+
+            using (ITestHost host = TestHelpers.GetJobHost(
+                this.loggerProvider,
+                nameof(this.DurableHttpAsync_Synchronous_RetryExceededWithHttpNotFoundViaEnsureSuccessException),
+                enableExtendedSessions: false,
+                storageProviderType: storageProvider,
+                durableHttpMessageHandler: new DurableHttpMessageHandlerFactory(httpMessageHandler)))
+            {
+                await host.StartAsync();
+
+                Dictionary<string, string> headers = new Dictionary<string, string>
+                {
+                    { "Accept", "application/json" },
+                };
+                TestDurableHttpRequest testRequest = new TestDurableHttpRequest(
+                    httpMethod: HttpMethod.Get,
+                    headers: headers,
+                    firstRetryInterval: TimeSpan.FromSeconds(1),
+                    maxNumberOfAttempts: 3)
+                {
+                    StatusCodesToRetry = new List<HttpStatusCode> { HttpStatusCode.NotFound },
+                };
+
+                string functionName = nameof(TestOrchestrations.CallHttpAsyncOrchestrator);
+                var client = await host.StartOrchestratorAsync(functionName, testRequest, this.output);
+                var status = await client.WaitForCompletionAsync(this.output, timeout: TimeSpan.FromSeconds(400));
+
+                var output = status?.Output;
+                Assert.Equal(3, mockSynchronousHttpMessageHandlerCount);
+                Assert.Equal(OrchestrationRuntimeStatus.Failed, status?.RuntimeStatus);
+
+                await host.StopAsync();
+            }
+        }
+
+        /// <summary>
+        /// End-to-end test which checks if the CallHttpAsync Orchestrator retries when the
+        /// HTTP request times out and the CallHttpAsync API throws a TimeoutException.
+        /// </summary>
+        [Theory]
+        [Trait("Category", PlatformSpecificHelpers.TestCategory)]
+        [MemberData(nameof(TestDataGenerator.GetFullFeaturedStorageProviderOptions), MemberType = typeof(TestDataGenerator))]
+        public async Task DurableHttpAsync_Synchronous_SuccessWithNoRetryWithHttpNotFoundIfNotSpecified(string storageProvider)
+        {
+            HttpMessageHandler httpMessageHandler = MockSynchronousHttpMessageHandlerWithHttp404();
+
+            using (ITestHost host = TestHelpers.GetJobHost(
+                this.loggerProvider,
+                nameof(this.DurableHttpAsync_Synchronous_SuccessWithNoRetryWithHttpNotFoundIfNotSpecified),
+                enableExtendedSessions: false,
+                storageProviderType: storageProvider,
+                durableHttpMessageHandler: new DurableHttpMessageHandlerFactory(httpMessageHandler)))
+            {
+                await host.StartAsync();
+
+                Dictionary<string, string> headers = new Dictionary<string, string>
+                {
+                    { "Accept", "application/json" },
+                };
+                TestDurableHttpRequest testRequest = new TestDurableHttpRequest(
+                    httpMethod: HttpMethod.Get,
+                    headers: headers,
+                    firstRetryInterval: TimeSpan.FromSeconds(1),
+                    maxNumberOfAttempts: 3)
+                {
+                    StatusCodesToRetry = new List<HttpStatusCode> { HttpStatusCode.Forbidden },
+                };
+
+                string functionName = nameof(TestOrchestrations.CallHttpAsyncOrchestrator);
+                var client = await host.StartOrchestratorAsync(functionName, testRequest, this.output);
+                var status = await client.WaitForCompletionAsync(this.output, timeout: TimeSpan.FromSeconds(400));
+
+                Assert.Equal(1, mockSynchronousHttpMessageHandlerCount);
+
+                var output = status?.Output;
+                DurableHttpResponse response = output.ToObject<DurableHttpResponse>();
+                Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
+
+                await host.StopAsync();
+            }
+        }
+
+        /// <summary>
         /// End-to-end test which checks if the CallHttpAsync Orchestrator fails  when the
         /// target url doesn't exist and throws an HttpRequestException.
         /// </summary>
@@ -393,6 +576,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
                 var status = await client.WaitForCompletionAsync(this.output);
 
                 var output = status?.Output;
+                Assert.Equal(1, mockSynchronousHttpMessageHandlerCount);
                 Assert.Contains("No such host is known.", output.ToString());
                 Assert.Equal(OrchestrationRuntimeStatus.Failed, status?.RuntimeStatus);
 
@@ -1549,15 +1733,44 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
 
         private static HttpMessageHandler MockSynchronousHttpMessageHandlerWithHttpRequestException()
         {
-            HttpResponseMessage httpResponseMessage = CreateTestHttpResponseMessage(HttpStatusCode.NotFound);
-
-            httpResponseMessage.Content = new ExceptionThrowingContent(new HttpRequestException("No such host is known."));
-
+            mockSynchronousHttpMessageHandlerCount = 0;
             var handlerMock = new Mock<HttpMessageHandler>(MockBehavior.Strict);
             handlerMock
                .Protected()
                .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
-               .ReturnsAsync(httpResponseMessage);
+               .ReturnsAsync(() =>
+               {
+                   // We create a new response every time because by the virtue of completing, the response object gets disposed
+                   // so if we reused the same object in our ReturnsAsync() an ObjectDisposedException gets thrown
+                   mockSynchronousHttpMessageHandlerCount++;
+
+                   HttpResponseMessage httpResponseMessage = CreateTestHttpResponseMessage(HttpStatusCode.NotFound);
+
+                   httpResponseMessage.Content = new ExceptionThrowingContent(new HttpRequestException("No such host is known."));
+
+                   return httpResponseMessage;
+               });
+
+            return handlerMock.Object;
+        }
+
+        private static HttpMessageHandler MockSynchronousHttpMessageHandlerWithHttp404()
+        {
+            mockSynchronousHttpMessageHandlerCount = 0;
+            var handlerMock = new Mock<HttpMessageHandler>(MockBehavior.Strict);
+            handlerMock
+               .Protected()
+               .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
+               .ReturnsAsync(() =>
+               {
+                   // We create a new response every time because by the virtue of completing, the response object gets disposed
+                   // so if we reused the same object in our ReturnsAsync() an ObjectDisposedException gets thrown
+                   mockSynchronousHttpMessageHandlerCount++;
+
+                   HttpResponseMessage httpResponseMessage = CreateTestHttpResponseMessage(HttpStatusCode.NotFound);
+
+                   return httpResponseMessage;
+               });
 
             return handlerMock.Object;
         }