+++ test/Common/TestDurableHttpRequest.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Net;
 using System.Net.Http;
 using System.Runtime.Serialization;
 using Newtonsoft.Json;
@@ -18,7 +19,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
     [DataContract]
     public class TestDurableHttpRequest
     {
-        public TestDurableHttpRequest(HttpMethod httpMethod, string uri = "https://www.dummy-url.com", IDictionary<string, string> headers = null, string content = null, ITokenSource tokenSource = null, TimeSpan? timeout = null)
+        public TestDurableHttpRequest(HttpMethod httpMethod, string uri = "https://www.dummy-url.com", IDictionary<string, string> headers = null, string content = null, ITokenSource tokenSource = null, TimeSpan? timeout = null, TimeSpan? firstRetryInterval = null, int? maxNumberOfAttempts = null)
         {
             this.HttpMethod = httpMethod;
             this.Uri = uri;
@@ -26,6 +27,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
             this.Content = content;
             this.TokenSource = tokenSource;
             this.Timeout = timeout;
+            this.FirstRetryInterval = firstRetryInterval;
+            this.MaxNumberOfAttempts = maxNumberOfAttempts;
         }
 
         [DataMember]
@@ -49,7 +52,16 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
         [DataMember]
         public bool AsynchronousPatternEnabled { get; set; } = true;
 
-        [DataMember]
+        [DataMember(EmitDefaultValue = false)]
         public TimeSpan? Timeout { get; set; }
+
+        [DataMember(EmitDefaultValue = false)]
+        public TimeSpan? FirstRetryInterval { get; set; }
+
+        [DataMember(EmitDefaultValue = false)]
+        public int? MaxNumberOfAttempts { get; set; }
+
+        [DataMember(EmitDefaultValue = false)]
+        public IList<HttpStatusCode> StatusCodesToRetry { get; set; }
     }
 }